+++ src/lib/openjp2/t1_opt.cpp
@@ -768,7 +768,7 @@ double t1_opt_encode_cblk(t1_opt_t *t1,
 				pass->rate--;
 			}
 		}
-		pass->len = pass->rate - (passno == 0 ? 0 : cblk->passes[passno - 1].rate);
+		pass->len = (uint16_t)(pass->rate - (passno == 0 ? 0 : cblk->passes[passno - 1].rate));
 		assert(pass->len != -1);
 	}
 	return cumwmsedec;
